tests/requirements.txt
coverage
pytest
pytest-cov
python-libsbml
pyneuroml
scipy
matplotlib
